extract SnapshotBundle
